네임스페이스 및 폴더 이름 변경
