+++ MoneyNoteUWP/Pages/HomePage.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="MoneyNote.Views.HomePage"
+    x:Class="MoneyNote.Pages.HomePage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:MoneyNote.Views"
+    xmlns:local="using:MoneyNote.Pages"
     xmlns:model="using:MoneyNoteLibrary.Models"
     xmlns:custom="using:MoneyNote.UserControls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
