+++ MoneyNoteUWP/Pages/MoneyBasicListPage.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="MoneyNote.Views.MoneyBasicListPage"
+    x:Class="MoneyNote.Pages.MoneyBasicListPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:MoneyNote.Views"
+    xmlns:local="using:MoneyNote.Pages"
     xmlns:model="using:MoneyNoteLibrary.Models"
     xmlns:viewmodel="using:MoneyNoteLibrary.ViewModels"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
