+++ MoneyNoteUWP/Pages/MoneyDetailView.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="MoneyNote.Views.MoneyDetailView"
+    x:Class="MoneyNote.Pages.MoneyDetailView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:MoneyNote.Views"
+    xmlns:local="using:MoneyNote.Pages"
     xmlns:model="using:MoneyNoteLibrary.Models"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
